first build from TypeScript

+++ managers/objectManager.js
@@ -1,12 +1,9 @@
-import 'three';
-
 class ObjectManager {
-    /**
-     * 
-     * @param {THREE.Scene} scene
-     * @param {THREE.GridHelper} grid
-     * @param {THREE.Mesh} plane
-     */
+    scene;
+    grid;
+    plane;
+    objects;
+    meshLookup;
     constructor(scene, grid, plane) {
         this.scene = scene;
         this.grid = grid;
@@ -14,22 +11,16 @@ class ObjectManager {
         this.objects = new Map();
         this.meshLookup = new Map();
     }
-
     addObject(visualObject) {
         const mesh = visualObject.getMesh();
+        if (!mesh) {
+            console.error(`addObject: Object with id ${visualObject.getUUID()} has no mesh!`);
+            return;
+        }
         this.scene.add(mesh);
         this.objects.set(visualObject.getUUID(), visualObject);
         this.meshLookup.set(mesh, visualObject);
     }
-
-    /*addEditObject(object, mesh){
-        object.userData.selectable = false;
-        object.userData.editable = true;
-        mesh.add(object);
-        const editObject = { object: object, mesh: mesh };
-        this.editObjects.push(editObject);
-    }*/
-
     getObjectByUUID(uuid) {
         if (!this.objects.has(uuid)) {
             console.error(`getObjectById: Object with uuid ${uuid} not found!`);
@@ -37,19 +28,15 @@ class ObjectManager {
         }
         return this.objects.get(uuid);
     }
-
-    selectable(mesh){
+    selectable(mesh) {
         return this.meshLookup.has(mesh);
     }
-
-    isEditHandle(mesh){
+    isEditHandle(mesh) {
         return mesh.userData.isEditHandle;
     }
-
-    isVisualObject(mesh){
-        return this.meshLookup.has(mesh);  
+    isVisualObject(mesh) {
+        return this.meshLookup.has(mesh);
     }
-
     getVisualObjectByMesh(mesh) {
         if (!this.meshLookup.has(mesh)) {
             console.error(`getVisualObjectByMesh: Object with mesh ${mesh} not found!`);
@@ -57,46 +44,25 @@ class ObjectManager {
         }
         return this.meshLookup.get(mesh);
     }
-
-    /**
-     * 
-     * @param {string} id
-     * @returns {void}
-     * @description Removes an object from the scene
-     */
     removeObject(uuid) {
-        const visualObject = this.getObjectById(uuid);
+        const visualObject = this.getObjectByUUID(uuid);
         if (!visualObject) {
             console.error(`removeObject: Object with id ${uuid} not found!`);
             return;
         }
-
         const mesh = visualObject.getMesh();
+        if (!mesh) {
+            console.error(`removeObject: Object with id ${uuid} has no mesh!`);
+            return;
+        }
         this.scene.remove(mesh);
         this.objects.delete(uuid);
         this.meshLookup.delete(mesh);
     }
-
-    /*removeEditObjects(){
-        this.editObjects.forEach(editObject => editObject.mesh.remove(editObject.object));
-        this.editObjects = [];
-    }*/
-
-    /**
-     * 
-     * @param {Mesh} object
-     * @returns {boolean} True if the object is the grid, false otherwise
-     */
-    isGrid(object){
+    isGrid(object) {
         return object === this.grid;
     }
-  
-    /**
-     * 
-     * @param {Mesh} object
-     * @returns {boolean} True if the object is the plane, false otherwise
-     */
-    isPlane(object){
+    isPlane(object) {
         return object === this.plane;
     }
 }
